mailhandler/requirements.txt
python-dotenv==1.0.0
exchangelib==5.4.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
tqdm==4.66.1
html2text==2020.1.16
msal==1.24.1
requests-oauthlib==1.3.1 